Implement useGeolocation. Make geolocation api call in Main component into the custom hook
